+++ packages/sysutils/v4l-utils/package.mk
@@ -5,26 +5,34 @@
 # with 1.0.0 repeat delay is broken. test on upgrade
 
 PKG_NAME="v4l-utils"
-PKG_VERSION="1.30.1"
-PKG_SHA256="c1cf549c2ec3cf39eb5ec7bf15731349e61b26a21b5e963922db422333bae197"
+PKG_VERSION="1.32.0"
+PKG_SHA256="6828828a17775526eb93fb258a9294d1d1073d633c344dd71ecd4e7a1ffb7dfc"
 PKG_LICENSE="GPL"
 PKG_SITE="https://linuxtv.org/"
 PKG_URL="https://linuxtv.org/downloads/v4l-utils/${PKG_NAME}-${PKG_VERSION}.tar.xz"
 
-[[ "${DEVICE}" != "Amlogic-old" ]] && PKG_DEPENDS_TARGET="libbpf"
-PKG_DEPENDS_TARGET="toolchain alsa-lib elfutils gcc-bpf:host ${PKG_DEPENDS_TARGET} systemd zlib"
+PKG_DEPENDS_TARGET="toolchain alsa-lib elfutils gcc-bpf:host systemd zlib"
+
+if [ "${DEVICE}" != "Amlogic-old" ]; then
+  PKG_DEPENDS_TARGET+=" libbpf"
+fi
 
 PKG_LONGDESC="Linux V4L2 and DVB API utilities and v4l libraries (libv4l)."
 PKG_STAMP_VAR="$IR_REMOTE_KEYMAPS"
 PKG_TOOLCHAIN="meson"
 
-# Flags traduzidas para manter o comportamento do 1.24.1
+if [ "${DEVICE}" = "Amlogic-old" ]; then
+  BPF_OPT="disabled"
+else
+  BPF_OPT="enabled"
+fi
+
 PKG_MESON_OPTS_TARGET="-Djpeg=disabled \
                        -Dgconv=disabled \
                        -Dqvidcap=disabled \
                        -Dqv4l2=disabled \
                        -Ddoxygen-doc=disabled \
-                       -Dbpf=$(if [ "${DEVICE}" = "Amlogic-old" ]; then echo "disabled"; else echo "enabled"; fi) \
+                       -Dbpf=${BPF_OPT} \
                        -Dudevdir=/usr/lib/udev"
 
 # Deixamos o EmuELEC gerenciar a pasta de build do Meson automaticamente
@@ -46,11 +54,11 @@ post_makeinstall_target() {
   local f keymap
 
   # Lógica original: Remove compliance se não for devel
-  if [ ! "${LIBREELEC_VERSION}" == "devel" ]; then
+  if [ ! "${DISTRO_VERSION}" == "devel" ]; then
     rm -f ${INSTALL}/usr/bin/v4l2-compliance
   fi
 
-  # Estrutura de links simbólicos original
+  rm -rf ${INSTALL}/usr/lib/gconv
   rm -rf ${INSTALL}/etc/rc_keymaps
     ln -sf /storage/.config/rc_keymaps ${INSTALL}/etc/rc_keymaps
 
@@ -92,7 +100,7 @@ EOF
   fi
 
   # Copia o contrib script (Lógica original)
-  cp ${PKG_BUILD}/contrib/lircd2toml.py ${INSTALL}/usr/bin/
+  [ -f ${PKG_BUILD}/contrib/lircd2toml.py ] && cp ${PKG_BUILD}/contrib/lircd2toml.py ${INSTALL}/usr/bin/
 
   # Limpeza NextOS (Mantendo o sistema Elite leve)
   rm -rf ${INSTALL}/usr/include
